requirements.txt
ollama
pgvector
psycopg2-binary
python-dotenv
matplotlib
scikit-learn
networkx 